@@ -9,56 +9,59 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20140206003351) do
+ActiveRecord::Schema.define(version: 20140211224225) do
 
-  create_table "active_admin_comments", :force => true do |t|
-    t.string   "resource_id",   :null => false
-    t.string   "resource_type", :null => false
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
+  create_table "active_admin_comments", force: true do |t|
+    t.string   "resource_id",   null: false
+    t.string   "resource_type", null: false
     t.integer  "author_id"
     t.string   "author_type"
     t.text     "body"
-    t.datetime "created_at",    :null => false
-    t.datetime "updated_at",    :null => false
+    t.datetime "created_at",    null: false
+    t.datetime "updated_at",    null: false
     t.string   "namespace"
   end
 
-  add_index "active_admin_comments", ["author_type", "author_id"], :name => "index_active_admin_comments_on_author_type_and_author_id"
-  add_index "active_admin_comments", ["namespace"], :name => "index_active_admin_comments_on_namespace"
-  add_index "active_admin_comments", ["resource_type", "resource_id"], :name => "index_admin_notes_on_resource_type_and_resource_id"
+  add_index "active_admin_comments", ["author_type", "author_id"], name: "index_active_admin_comments_on_author_type_and_author_id", using: :btree
+  add_index "active_admin_comments", ["namespace"], name: "index_active_admin_comments_on_namespace", using: :btree
+  add_index "active_admin_comments", ["resource_type", "resource_id"], name: "index_admin_notes_on_resource_type_and_resource_id", using: :btree
 
-  create_table "activities", :force => true do |t|
+  create_table "activities", force: true do |t|
     t.integer  "user_id"
     t.string   "action"
     t.integer  "trackable_id"
     t.string   "trackable_type"
-    t.datetime "created_at",     :null => false
-    t.datetime "updated_at",     :null => false
+    t.datetime "created_at",     null: false
+    t.datetime "updated_at",     null: false
   end
 
-  add_index "activities", ["trackable_id"], :name => "index_activities_on_trackable_id"
-  add_index "activities", ["user_id"], :name => "index_activities_on_user_id"
+  add_index "activities", ["trackable_id"], name: "index_activities_on_trackable_id", using: :btree
+  add_index "activities", ["user_id"], name: "index_activities_on_user_id", using: :btree
 
-  create_table "admin_users", :force => true do |t|
-    t.string   "email",                  :default => "", :null => false
-    t.string   "encrypted_password",     :default => "", :null => false
+  create_table "admin_users", force: true do |t|
+    t.string   "email",                  default: "", null: false
+    t.string   "encrypted_password",     default: "", null: false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          :default => 0
+    t.integer  "sign_in_count",          default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
-    t.datetime "created_at",                             :null => false
-    t.datetime "updated_at",                             :null => false
+    t.datetime "created_at",                          null: false
+    t.datetime "updated_at",                          null: false
   end
 
-  add_index "admin_users", ["email"], :name => "index_admin_users_on_email", :unique => true
-  add_index "admin_users", ["reset_password_token"], :name => "index_admin_users_on_reset_password_token", :unique => true
+  add_index "admin_users", ["email"], name: "index_admin_users_on_email", unique: true, using: :btree
+  add_index "admin_users", ["reset_password_token"], name: "index_admin_users_on_reset_password_token", unique: true, using: :btree
 
-  create_table "apps", :force => true do |t|
+  create_table "apps", force: true do |t|
     t.string   "first"
     t.string   "last"
     t.string   "email"
@@ -76,25 +79,25 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.string   "marketing"
     t.string   "technical"
     t.string   "city"
-    t.datetime "created_at",         :null => false
-    t.datetime "updated_at",         :null => false
+    t.datetime "created_at",         null: false
+    t.datetime "updated_at",         null: false
     t.string   "photo_file_name"
     t.string   "photo_content_type"
     t.integer  "photo_file_size"
     t.datetime "photo_updated_at"
   end
 
-  create_table "assignments", :force => true do |t|
+  create_table "assignments", force: true do |t|
     t.string   "title"
     t.text     "description"
     t.string   "vimeo_url"
     t.string   "preview_url"
-    t.datetime "created_at",                                  :null => false
-    t.datetime "updated_at",                                  :null => false
+    t.datetime "created_at",                               null: false
+    t.datetime "updated_at",                               null: false
     t.integer  "course_id"
-    t.integer  "order_id",                 :default => 0
+    t.integer  "order_id",                 default: 0
     t.integer  "survey_id"
-    t.boolean  "require_upload",           :default => false
+    t.boolean  "require_upload",           default: false
     t.string   "speaker_name"
     t.text     "speaker_bio"
     t.string   "speaker_linkedin"
@@ -110,40 +113,40 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.integer  "user_assignment_id"
   end
 
-  add_index "assignments", ["slug"], :name => "index_assignments_on_slug"
+  add_index "assignments", ["slug"], name: "index_assignments_on_slug", using: :btree
 
-  create_table "attachments", :force => true do |t|
+  create_table "attachments", force: true do |t|
     t.string   "url"
     t.string   "type"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "authorships", :force => true do |t|
+  create_table "authorships", force: true do |t|
     t.integer  "user_id"
     t.integer  "skill_id"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  add_index "authorships", ["skill_id"], :name => "index_authorships_on_skill_id"
-  add_index "authorships", ["user_id"], :name => "index_authorships_on_user_id"
+  add_index "authorships", ["skill_id"], name: "index_authorships_on_skill_id", using: :btree
+  add_index "authorships", ["user_id"], name: "index_authorships_on_user_id", using: :btree
 
-  create_table "badges", :force => true do |t|
+  create_table "badges", force: true do |t|
     t.string   "name"
     t.integer  "user_id"
     t.integer  "course_id"
-    t.datetime "created_at",        :null => false
-    t.datetime "updated_at",        :null => false
+    t.datetime "created_at",        null: false
+    t.datetime "updated_at",        null: false
     t.string   "icon_file_name"
     t.string   "icon_content_type"
     t.integer  "icon_file_size"
     t.datetime "icon_updated_at"
   end
 
-  create_table "bucketlist_items", :force => true do |t|
-    t.datetime "created_at",    :null => false
-    t.datetime "updated_at",    :null => false
+  create_table "bucketlist_items", force: true do |t|
+    t.datetime "created_at",    null: false
+    t.datetime "updated_at",    null: false
     t.integer  "user_id"
     t.integer  "bucketlist_id"
     t.string   "title"
@@ -151,17 +154,17 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.boolean  "complete"
   end
 
-  create_table "bucketlists", :force => true do |t|
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+  create_table "bucketlists", force: true do |t|
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.integer  "user_id"
   end
 
-  create_table "courses", :force => true do |t|
+  create_table "courses", force: true do |t|
     t.string   "title"
     t.text     "description"
-    t.datetime "created_at",                    :null => false
-    t.datetime "updated_at",                    :null => false
+    t.datetime "created_at",                    null: false
+    t.datetime "updated_at",                    null: false
     t.string   "course_icon_file_name"
     t.string   "course_icon_content_type"
     t.integer  "course_icon_file_size"
@@ -179,51 +182,54 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.string   "tile_image_content_type"
     t.integer  "tile_image_file_size"
     t.datetime "tile_image_updated_at"
+    t.string   "slug"
   end
 
-  create_table "delayed_jobs", :force => true do |t|
-    t.integer  "priority",   :default => 0, :null => false
-    t.integer  "attempts",   :default => 0, :null => false
-    t.text     "handler",                   :null => false
+  add_index "courses", ["slug"], name: "index_courses_on_slug", unique: true, using: :btree
+
+  create_table "delayed_jobs", force: true do |t|
+    t.integer  "priority",   default: 0, null: false
+    t.integer  "attempts",   default: 0, null: false
+    t.text     "handler",                null: false
     t.text     "last_error"
     t.datetime "run_at"
     t.datetime "locked_at"
     t.datetime "failed_at"
     t.string   "locked_by"
     t.string   "queue"
-    t.datetime "created_at",                :null => false
-    t.datetime "updated_at",                :null => false
+    t.datetime "created_at",             null: false
+    t.datetime "updated_at",             null: false
   end
 
-  add_index "delayed_jobs", ["priority", "run_at"], :name => "delayed_jobs_priority"
+  add_index "delayed_jobs", ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
 
-  create_table "events", :force => true do |t|
+  create_table "events", force: true do |t|
     t.string   "name"
     t.datetime "start_time"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at",  null: false
+    t.datetime "updated_at",  null: false
     t.text     "description"
     t.string   "location"
     t.integer  "user_id"
   end
 
-  add_index "events", ["user_id"], :name => "index_events_on_user_id"
+  add_index "events", ["user_id"], name: "index_events_on_user_id", using: :btree
 
-  create_table "feedbacks", :force => true do |t|
+  create_table "feedbacks", force: true do |t|
     t.string   "type"
     t.text     "content"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "messages", :force => true do |t|
+  create_table "messages", force: true do |t|
     t.text     "content"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.integer  "user_id"
   end
 
-  create_table "milestones", :force => true do |t|
+  create_table "milestones", force: true do |t|
     t.text     "vision"
     t.text     "creativity"
     t.text     "speedstrength"
@@ -231,22 +237,22 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.text     "power"
     t.text     "survival"
     t.text     "brilliance"
-    t.datetime "created_at",    :null => false
-    t.datetime "updated_at",    :null => false
+    t.datetime "created_at",    null: false
+    t.datetime "updated_at",    null: false
   end
 
-  create_table "payments", :force => true do |t|
+  create_table "payments", force: true do |t|
     t.string   "name"
     t.text     "email"
     t.float    "fee"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "posts", :force => true do |t|
+  create_table "posts", force: true do |t|
     t.text     "content"
-    t.datetime "created_at",        :null => false
-    t.datetime "updated_at",        :null => false
+    t.datetime "created_at",        null: false
+    t.datetime "updated_at",        null: false
     t.string   "file_file_name"
     t.string   "file_content_type"
     t.integer  "file_file_size"
@@ -254,35 +260,35 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.integer  "user_id"
   end
 
-  create_table "resources", :force => true do |t|
+  create_table "resources", force: true do |t|
     t.string   "title"
     t.string   "url"
     t.text     "description"
     t.string   "category"
-    t.datetime "created_at",         :null => false
-    t.datetime "updated_at",         :null => false
+    t.datetime "created_at",         null: false
+    t.datetime "updated_at",         null: false
     t.string   "image_file_name"
     t.string   "image_content_type"
     t.integer  "image_file_size"
     t.datetime "image_updated_at"
   end
 
-  create_table "skills", :force => true do |t|
+  create_table "skills", force: true do |t|
     t.string   "name"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "survey_answers", :force => true do |t|
+  create_table "survey_answers", force: true do |t|
     t.integer  "attempt_id"
     t.integer  "question_id"
     t.integer  "option_id"
     t.boolean  "correct"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at",  null: false
+    t.datetime "updated_at",  null: false
   end
 
-  create_table "survey_attempts", :force => true do |t|
+  create_table "survey_attempts", force: true do |t|
     t.integer "participant_id"
     t.string  "participant_type"
     t.integer "survey_id"
@@ -290,53 +296,53 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.integer "score"
   end
 
-  create_table "survey_options", :force => true do |t|
+  create_table "survey_options", force: true do |t|
     t.integer  "question_id"
-    t.integer  "weight",      :default => 0
+    t.integer  "weight",      default: 0
     t.string   "text"
     t.boolean  "correct"
-    t.datetime "created_at",                 :null => false
-    t.datetime "updated_at",                 :null => false
+    t.datetime "created_at",              null: false
+    t.datetime "updated_at",              null: false
   end
 
-  create_table "survey_questions", :force => true do |t|
+  create_table "survey_questions", force: true do |t|
     t.integer  "survey_id"
     t.string   "text"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "survey_surveys", :force => true do |t|
+  create_table "survey_surveys", force: true do |t|
     t.string   "name"
     t.text     "description"
-    t.integer  "attempts_number", :default => 0
-    t.boolean  "finished",        :default => false
-    t.boolean  "active",          :default => false
-    t.datetime "created_at",                         :null => false
-    t.datetime "updated_at",                         :null => false
+    t.integer  "attempts_number", default: 0
+    t.boolean  "finished",        default: false
+    t.boolean  "active",          default: false
+    t.datetime "created_at",                      null: false
+    t.datetime "updated_at",                      null: false
   end
 
-  create_table "taggings", :force => true do |t|
+  create_table "taggings", force: true do |t|
     t.integer  "tag_id"
     t.integer  "taggable_id"
     t.string   "taggable_type"
     t.integer  "tagger_id"
     t.string   "tagger_type"
-    t.string   "context",       :limit => 128
+    t.string   "context",       limit: 128
     t.datetime "created_at"
   end
 
-  add_index "taggings", ["tag_id"], :name => "index_taggings_on_tag_id"
-  add_index "taggings", ["taggable_id", "taggable_type", "context"], :name => "index_taggings_on_taggable_id_and_taggable_type_and_context"
+  add_index "taggings", ["tag_id"], name: "index_taggings_on_tag_id", using: :btree
+  add_index "taggings", ["taggable_id", "taggable_type", "context"], name: "index_taggings_on_taggable_id_and_taggable_type_and_context", using: :btree
 
-  create_table "tags", :force => true do |t|
+  create_table "tags", force: true do |t|
     t.string "name"
   end
 
-  create_table "user_assignments", :force => true do |t|
+  create_table "user_assignments", force: true do |t|
     t.text     "text"
-    t.datetime "created_at",          :null => false
-    t.datetime "updated_at",          :null => false
+    t.datetime "created_at",          null: false
+    t.datetime "updated_at",          null: false
     t.string   "upload_file_name"
     t.string   "upload_content_type"
     t.integer  "upload_file_size"
@@ -346,18 +352,18 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.string   "link"
   end
 
-  create_table "user_comments", :force => true do |t|
+  create_table "user_comments", force: true do |t|
     t.text     "content"
     t.integer  "commentable_id"
     t.string   "commentable_type"
     t.integer  "user_id"
-    t.datetime "created_at",       :null => false
-    t.datetime "updated_at",       :null => false
+    t.datetime "created_at",       null: false
+    t.datetime "updated_at",       null: false
   end
 
-  add_index "user_comments", ["commentable_id", "commentable_type"], :name => "index_comments_on_commentable_id_and_commentable_type"
+  add_index "user_comments", ["commentable_id", "commentable_type"], name: "index_comments_on_commentable_id_and_commentable_type", using: :btree
 
-  create_table "users", :force => true do |t|
+  create_table "users", force: true do |t|
     t.string   "first_name"
     t.string   "last_name"
     t.text     "bio"
@@ -370,14 +376,14 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.string   "state"
     t.string   "country"
     t.string   "zip"
-    t.datetime "created_at",                                      :null => false
-    t.datetime "updated_at",                                      :null => false
-    t.string   "email",                  :default => "",          :null => false
-    t.string   "encrypted_password",     :default => "",          :null => false
+    t.datetime "created_at",                                   null: false
+    t.datetime "updated_at",                                   null: false
+    t.string   "email",                  default: "",          null: false
+    t.string   "encrypted_password",     default: "",          null: false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          :default => 0
+    t.integer  "sign_in_count",          default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
@@ -386,13 +392,13 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.string   "avatar_content_type"
     t.integer  "avatar_file_size"
     t.datetime "avatar_updated_at"
-    t.string   "online",                 :default => "online"
+    t.string   "online",                 default: "online"
     t.string   "team"
     t.string   "skype"
     t.string   "gmail"
     t.string   "slug"
-    t.float    "latitude",               :default => 37.5638
-    t.float    "longitude",              :default => -122.325192
+    t.float    "latitude",               default: 37.5638
+    t.float    "longitude",              default: -122.325192
     t.string   "employment"
     t.string   "instagram"
     t.string   "angellist"
@@ -401,11 +407,11 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.integer  "nCounter"
   end
 
-  add_index "users", ["email"], :name => "index_users_on_email", :unique => true
-  add_index "users", ["reset_password_token"], :name => "index_users_on_reset_password_token", :unique => true
-  add_index "users", ["slug"], :name => "index_users_on_slug"
+  add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
+  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
+  add_index "users", ["slug"], name: "index_users_on_slug", using: :btree
 
-  create_table "videos", :force => true do |t|
+  create_table "videos", force: true do |t|
     t.string   "title"
     t.text     "description"
     t.string   "vimeo_url"
@@ -418,8 +424,8 @@ ActiveRecord::Schema.define(:version => 20140206003351) do
     t.string   "speaker_angel"
     t.string   "category"
     t.string   "slug"
-    t.datetime "created_at",               :null => false
-    t.datetime "updated_at",               :null => false
+    t.datetime "created_at",               null: false
+    t.datetime "updated_at",               null: false
     t.string   "speaker_pic_file_name"
     t.string   "speaker_pic_content_type"
     t.integer  "speaker_pic_file_size"
