@@ -33,7 +33,7 @@ ActiveAdmin.register Course do
       :storage => :s3,
       :s3_credentials => "#{Rails.root}/config/s3.yml"
     end
-  f.buttons
+  f.actions
  end
 
 end
