@@ -43,7 +43,7 @@ ActiveAdmin.register User do
       :storage => :s3,
       :s3_credentials => "#{Rails.root}/config/s3.yml"
    end
-  f.buttons
+  f.actions
  end
 
 end
