@@ -28,7 +28,7 @@ ActiveAdmin.register Video do
       :storage => :s3,
       :s3_credentials => "#{Rails.root}/config/s3.yml"
    end
-  f.buttons
+  f.actions
  end
 
 end
