@@ -1,5 +1,5 @@
 class ApplicationController < ActionController::Base
-
+ 
   protect_from_forgery
   before_filter :cors_preflight_check
   after_filter :cors_set_access_control_headers
@@ -26,8 +26,16 @@ end
   #PRODUCTION FAYE: : http://fsrvrproduction.herokuapp.com
   #STAGING FAYE:  http://fayesrvr.herokuapp.com
 
+ # before_filter :miniprofiler
+  
+ #  private
+ #  def miniprofiler
+ #    Rack::MiniProfiler.authorize_request
+ #  end
+
   private
      def track_activity(trackable, action = params[:action])
        current_user.activities.create! action: action, trackable: trackable
      end  
 end
+
