@@ -18,7 +18,7 @@ class ProfilesController < ApplicationController
   end
 
   def profile
-    @user = User.find(params[:id])
+    @user = User.friendly.find(params[:id])
     @courses = Course.all
     @assignments = Assignment.all
     #@assignment = Assignment.find(params[:id])
