@@ -3,7 +3,7 @@ class UserAssignmentsController < ApplicationController
   load_and_authorize_resource
   
   def show
-    @user_assignment = UserAssignment.find(params[:id])
+    @user_assignment = UserAssignment.friendly.find(params[:id])
     @commentable = @user_assignment
     @comments = @commentable.user_comments.order(:created_at)
     @comment = UserComment.new
