@@ -6,8 +6,7 @@ OnlineSchool::Application.configure do
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
+  config.eager_load = true
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
@@ -19,23 +18,10 @@ OnlineSchool::Application.configure do
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :log
 
-  # Only use best-standards-support built into browsers
-  config.action_dispatch.best_standards_support = :builtin
-
-  # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
-
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
-
-  # Do not compress assets
-  config.assets.compress = false
-
   # Expands the lines which load the assets
-  config.assets.debug = true
+  config.assets.debug = false
   config.action_mailer.default_url_options = { :host => 'localhost' }
-    config.action_mailer.smtp_settings = {
+  config.action_mailer.smtp_settings = {
     :address   => "smtp.mandrillapp.com",
     :port      => 25, # ports 587 and 2525 are also supported with STARTTLS
     :enable_starttls_auto => true, # detects and uses STARTTLS
@@ -57,7 +43,7 @@ OnlineSchool::Application.configure do
 
   Paperclip.options[:command_path] = "/usr/local/bin/"
 
- # WS.config(:access_key_id => 'AWS_ACCESS_KEY_ID', :secret_access_key => 'AWS_SECRET_ACCESS_KEY')
+  # WS.config(:access_key_id => 'AWS_ACCESS_KEY_ID', :secret_access_key => 'AWS_SECRET_ACCESS_KEY')
 
   config.paperclip_defaults = {
     :storage => :s3,
@@ -67,6 +53,6 @@ OnlineSchool::Application.configure do
     }
   }
 
-  
+
 
 end
