@@ -3,7 +3,7 @@ OnlineSchool::Application.configure do
 
   # Code is not reloaded between requests
   config.cache_classes = true
-
+  config.eager_load = true
   # Full error reports are disabled and caching is turned on
   config.consider_all_requests_local       = false
   config.action_controller.perform_caching = true
@@ -12,7 +12,7 @@ OnlineSchool::Application.configure do
   config.serve_static_assets = false
 
   # Compress JavaScripts and CSS
-  config.assets.compress = true
+  config.assets.js_compress = :uglifier
 
   # Don't fallback to assets pipeline if a precompiled asset is missed
   config.assets.compile = true
